+++ packages/maskbook/src/plugins/COTM/hooks/useAllTokensOfOwner.ts
@@ -1,7 +1,7 @@
 import { useERC721TokenIdsOfOwner } from '../../../web3/hooks/useERC721TokensOfOwner'
 import type { ERC721TokenDetailed } from '../../../web3/types'
 
-export function useAllCOTM_TokensOfOwner(token?: ERC721TokenDetailed) {
+export function useAllTokensOfOwner(token?: ERC721TokenDetailed) {
     const { value: tokenIds, ...result } = useERC721TokenIdsOfOwner(token)
     return {
         ...result,
