+++ packages/maskbook/src/plugins/COTM/hooks/useTokensOfOwner.ts
@@ -1,10 +1,10 @@
 import { useERC721TokenIdsOfOwner } from '../../../web3/hooks/useERC721TokensOfOwner'
 import type { ERC721TokenDetailed } from '../../../web3/types'
-import { useCOTM_Tokens } from './useCOTM_Tokens'
+import { useTokens } from './useTokens'
 
-export function useCOTM_TokensOfOwner(token?: ERC721TokenDetailed) {
+export function useTokensOfOwner(token?: ERC721TokenDetailed) {
     const { value: tokenIds, ...result } = useERC721TokenIdsOfOwner(token)
-    const totalTokens = useCOTM_Tokens(token)
+    const totalTokens = useTokens(token)
 
     if (!token || !totalTokens.length)
         return {
