--- packages/maskbook/src/plugins/COTM/pipes.ts
@@ -1,1 +1,0 @@
-export function foo() {}
