+++ packages/maskbook/src/plugins/PluginSerivce.ts
@@ -7,5 +7,4 @@ import './Polls/utils'
 import './Transak/messages'
 import './Trader/messages'
 import './Polls/utils'
-import './Election2020/helpers'
-import './COTM/messages'
+import './Election2020/messages'
